Finalized DailyStreakVisualization component
