+++ src/components/DailyStreakVisualization/DailyStreakVisualization.jsx
@@ -1,4 +1,8 @@
-import React from 'react';
+import React, {
+  useEffect,
+  useRef,
+  useState,
+} from 'react';
 
 import "./DailyStreakVisualization.scss";
 
@@ -9,190 +13,172 @@ import findKeyWithLargestArray from '../../utils/findKeyWithLargestArray';
 
 const weekdays = [" ", ...WEEKDAY_STRINGS];
 
-// TODO: REMOVE, TEMP DATA
-const TEMP_GAME_RECORDS = [
-  {
-    completed: true,
-    lastPlayed: { seconds: 1688462821, nanoseconds: 0 },
-    gameTimer: 214500,
-    testGameRecord: true,
-    moves: 55,
-  },
-  {
-    completed: true,
-    lastPlayed: { seconds: 1688384400, nanoseconds: 0 },
-    gameTimer: 289000,
-    testGameRecord: true,
-    moves: 76,
-  },
-  {
-    completed: true,
-    lastPlayed: { seconds: 1688380800, nanoseconds: 0 },
-    gameTimer: 191000,
-    testGameRecord: true,
-    moves: 42,
-  },
-  {
-    completed: true,
-    lastPlayed: { seconds: 1688377200, nanoseconds: 0 },
-    gameTimer: 339000,
-    testGameRecord: true,
-    moves: 68,
-  },
-  {
-    completed: true,
-    lastPlayed: { seconds: 1688313600, nanoseconds: 0 },
-    gameTimer: 150000,
-    testGameRecord: true,
-    moves: 33,
-  },
-  {
-    completed: true,
-    lastPlayed: { seconds: 1688197200, nanoseconds: 0 },
-    gameTimer: 346000,
-    testGameRecord: true,
-    moves: 80,
-  },
-  {
-    completed: true,
-    lastPlayed: { seconds: 1688190000, nanoseconds: 0 },
-    gameTimer: 173000,
-    testGameRecord: true,
-    moves: 58,
-  },
-  {
-    completed: true,
-    lastPlayed: { seconds: 1687981200, nanoseconds: 0 },
-    gameTimer: 212000,
-    testGameRecord: true,
-    moves: 49,
-  },
-  {
-    completed: true,
-    lastPlayed: { seconds: 1687472400, nanoseconds: 0 },
-    gameTimer: 243000,
-    testGameRecord: true,
-    moves: 63,
-  },
-  {
-    completed: true,
-    lastPlayed: { seconds: 1687081200, nanoseconds: 0 },
-    gameTimer: 116000,
-    testGameRecord: true,
-    moves: 39,
-  },
-  {
-    completed: true,
-    lastPlayed: { seconds: 1686882000, nanoseconds: 0 },
-    gameTimer: 304000,
-    testGameRecord: true,
-    moves: 72,
-  },
-  {
-    completed: true,
-    lastPlayed: { seconds: 1686878400, nanoseconds: 0 },
-    gameTimer: 233000,
-    testGameRecord: true,
-    moves: 61,
-  },
-  {
-    completed: true,
-    lastPlayed: { seconds: 1686874800, nanoseconds: 0 },
-    gameTimer: 398000,
-    testGameRecord: true,
-    moves: 82,
-  },
-  {
-    completed: true,
-    lastPlayed: { seconds: 1686871200, nanoseconds: 0 },
-    gameTimer: 221000,
-    testGameRecord: true,
-    moves: 59,
-  },
-  {
-    completed: true,
-    lastPlayed: { seconds: 1686788400, nanoseconds: 0 },
-    gameTimer: 278000,
-    testGameRecord: true,
-    moves: 77,
-  },
-  {
-    completed: true,
-    lastPlayed: { seconds: 1686692400, nanoseconds: 0 },
-    gameTimer: 186000,
-    testGameRecord: true,
-    moves: 40,
-  },
-  {
-    completed: true,
-    lastPlayed: { seconds: 1686596400, nanoseconds: 0 },
-    gameTimer: 317000,
-    testGameRecord: true,
-    moves: 70,
-  },
-  {
-    completed: true,
-    lastPlayed: { seconds: 1686500400, nanoseconds: 0 },
-    gameTimer: 222000,
-    testGameRecord: true,
-    moves: 56,
-  },
-  {
-    completed: true,
-    lastPlayed: { seconds: 1686010800, nanoseconds: 0 },
-    gameTimer: 130000,
-    testGameRecord: true,
-    moves: 31,
-  },
-  {
-    completed: true,
-    lastPlayed: { seconds: 1684234800, nanoseconds: 0 },
-    gameTimer: 139800,
-    testGameRecord: true,
-    moves: 38,
-  },
-];
+// // TODO: REMOVE, TEMP DATA
+// const TEMP_GAME_RECORDS = [
+//   {
+//     completed: true,
+//     lastPlayed: { seconds: 1688462821, nanoseconds: 0 },
+//     gameTimer: 214500,
+//     testGameRecord: true,
+//     moves: 55,
+//   },
+//   {
+//     completed: true,
+//     lastPlayed: { seconds: 1688384400, nanoseconds: 0 },
+//     gameTimer: 289000,
+//     testGameRecord: true,
+//     moves: 76,
+//   },
+//   {
+//     completed: true,
+//     lastPlayed: { seconds: 1688380800, nanoseconds: 0 },
+//     gameTimer: 191000,
+//     testGameRecord: true,
+//     moves: 42,
+//   },
+//   {
+//     completed: true,
+//     lastPlayed: { seconds: 1688377200, nanoseconds: 0 },
+//     gameTimer: 339000,
+//     testGameRecord: true,
+//     moves: 68,
+//   },
+//   {
+//     completed: true,
+//     lastPlayed: { seconds: 1688313600, nanoseconds: 0 },
+//     gameTimer: 150000,
+//     testGameRecord: true,
+//     moves: 33,
+//   },
+//   {
+//     completed: true,
+//     lastPlayed: { seconds: 1688197200, nanoseconds: 0 },
+//     gameTimer: 346000,
+//     testGameRecord: true,
+//     moves: 80,
+//   },
+//   {
+//     completed: true,
+//     lastPlayed: { seconds: 1688190000, nanoseconds: 0 },
+//     gameTimer: 173000,
+//     testGameRecord: true,
+//     moves: 58,
+//   },
+//   {
+//     completed: true,
+//     lastPlayed: { seconds: 1687981200, nanoseconds: 0 },
+//     gameTimer: 212000,
+//     testGameRecord: true,
+//     moves: 49,
+//   },
+//   {
+//     completed: true,
+//     lastPlayed: { seconds: 1687472400, nanoseconds: 0 },
+//     gameTimer: 243000,
+//     testGameRecord: true,
+//     moves: 63,
+//   },
+//   {
+//     completed: true,
+//     lastPlayed: { seconds: 1687081200, nanoseconds: 0 },
+//     gameTimer: 116000,
+//     testGameRecord: true,
+//     moves: 39,
+//   },
+//   {
+//     completed: true,
+//     lastPlayed: { seconds: 1686882000, nanoseconds: 0 },
+//     gameTimer: 304000,
+//     testGameRecord: true,
+//     moves: 72,
+//   },
+//   {
+//     completed: true,
+//     lastPlayed: { seconds: 1686878400, nanoseconds: 0 },
+//     gameTimer: 233000,
+//     testGameRecord: true,
+//     moves: 61,
+//   },
+//   {
+//     completed: true,
+//     lastPlayed: { seconds: 1686874800, nanoseconds: 0 },
+//     gameTimer: 398000,
+//     testGameRecord: true,
+//     moves: 82,
+//   },
+//   {
+//     completed: true,
+//     lastPlayed: { seconds: 1686871200, nanoseconds: 0 },
+//     gameTimer: 221000,
+//     testGameRecord: true,
+//     moves: 59,
+//   },
+//   {
+//     completed: true,
+//     lastPlayed: { seconds: 1686788400, nanoseconds: 0 },
+//     gameTimer: 278000,
+//     testGameRecord: true,
+//     moves: 77,
+//   },
+//   {
+//     completed: true,
+//     lastPlayed: { seconds: 1686692400, nanoseconds: 0 },
+//     gameTimer: 186000,
+//     testGameRecord: true,
+//     moves: 40,
+//   },
+//   {
+//     completed: true,
+//     lastPlayed: { seconds: 1686596400, nanoseconds: 0 },
+//     gameTimer: 317000,
+//     testGameRecord: true,
+//     moves: 70,
+//   },
+//   {
+//     completed: true,
+//     lastPlayed: { seconds: 1686500400, nanoseconds: 0 },
+//     gameTimer: 222000,
+//     testGameRecord: true,
+//     moves: 56,
+//   },
+//   {
+//     completed: true,
+//     lastPlayed: { seconds: 1686010800, nanoseconds: 0 },
+//     gameTimer: 130000,
+//     testGameRecord: true,
+//     moves: 31,
+//   },
+//   {
+//     completed: true,
+//     lastPlayed: { seconds: 1684234800, nanoseconds: 0 },
+//     gameTimer: 139800,
+//     testGameRecord: true,
+//     moves: 38,
+//   },
+// ];
 
-// TODO: Also temp, but I might reuse this name later on
-//  this should also be a prop
-const gameRecords = TEMP_GAME_RECORDS.map((record) => {
-  record.lastPlayed = convertFromFirestoreTimestampToDate(record.lastPlayed);
-  return record;
-});
+// // TODO: Also temp, but I might reuse this name later on
+// //  this should also be a prop
+// const gameRecords = TEMP_GAME_RECORDS.map((record) => {
+//   record.lastPlayed = convertFromFirestoreTimestampToDate(record.lastPlayed);
+//   return record;
+// });
 
 // TODO: Move this logic - it should be within the body of the component 
 // BEGIN LOGIC
-const gameRecordsByDay = {};
+// const gameRecordsByDay = {};
 
-gameRecords.forEach((record) => {
-  if (!record.completed) {
-    return;
-  }
-
-  const { lastPlayed } = record;
-  const month = lastPlayed.getMonth() + 1;
-  const day = lastPlayed.getDate();
-  const dayKey = `${month}/${day}`;
-
-  if (gameRecordsByDay[dayKey]) {
-    gameRecordsByDay[dayKey].push(record);
-  }
-  else {
-    gameRecordsByDay[dayKey] = [record];
-  }
-});
-
-const largestKey = findKeyWithLargestArray(gameRecordsByDay);
-const dailyMax = gameRecordsByDay[largestKey].length;
-
-console.log("DSV: gameRecords:", gameRecords);
-console.log("DSV: gameRecordsByDay:", gameRecordsByDay);
-console.log("DSV: largestKey:", largestKey);
-console.log("DSV: dailyMax:", dailyMax);
 
 // END OF LOGIC TO MOVE INTO COMPONENT
 
-const DailyStreakVisualization = () => {
+const DailyStreakVisualization = ({
+  gameRecords,
+}) => {
+  const [ gameRecordsByDay, setGameRecordsByDay ] = useState();
+  const [ dailyMax, setDailyMax ] = useState();
+  
+  const largestKey = useRef(null);
+
   // TODO: only necessary for demo stuff?
   const dailyStreakCellNames = [
     "daily-streak-cell-00",
@@ -210,6 +196,29 @@ const DailyStreakVisualization = () => {
 
   const getRandomClassName = () => 
     dailyStreakCellNames[Math.floor(Math.random() * dailyStreakCellNames.length)];
+
+  const getCellClassName = (cellDate) => {
+    if (!cellDate || !dailyMax) {
+      return;
+    }
+
+    console.log(`DSV: getCellClassName: For ${cellDate}, we have ${getPuzzleCountOnDay(cellDate)}`);
+    let classNameIndex = Math.floor((getPuzzleCountOnDay(cellDate) / dailyMax) * 10);
+    const cellClassName = dailyStreakCellNames[classNameIndex];
+    console.log(`DSV: getCellClassName: classNameIndex is ${classNameIndex}`);
+    console.log(`DSV: getCellClassName: className is ${cellClassName}`);
+    
+    return cellClassName;
+  };
+
+  // helper function to get puzzle count on a given day
+  const getPuzzleCountOnDay = (cellDate) => {
+    if (!cellDate || !gameRecordsByDay || !gameRecordsByDay[cellDate] || !Array.isArray(gameRecordsByDay[cellDate])) {
+      return 0;
+    }
+
+    return gameRecordsByDay[cellDate].length;
+  };
 
   // TODO: Maybe move to its own module in utils?
   const getPastDates = (
@@ -260,57 +269,100 @@ const DailyStreakVisualization = () => {
 
   const cellWeeks = getPastDates();
 
+  useEffect(() => {
+    if (!gameRecords) {
+      return;
+    }
+
+    console.log("DSV: useEffect: Received gameRecords as prop:", gameRecords);
+
+    const newGameRecordsByDay = {};
+
+    // format gameRecords into gameRecordsByDay
+    gameRecords.forEach((record) => {
+      if (!record.completed) {
+        return;
+      }
+    
+      const { lastPlayed } = record;
+      const month = lastPlayed.getMonth() + 1;
+      const day = lastPlayed.getDate();
+      const dayKey = `${month}/${day}`;
+    
+      if (newGameRecordsByDay[dayKey]) {
+        newGameRecordsByDay[dayKey].push(record);
+      }
+      else {
+        newGameRecordsByDay[dayKey] = [record];
+      }
+    });
+
+    // find the largestKey and the dailyMax (most puzzles in one day)
+    
+    // TODO: Recalculated on memo?
+    largestKey.current = findKeyWithLargestArray(newGameRecordsByDay);
+    const newDailyMax = newGameRecordsByDay[largestKey.current].length;
+    
+    console.log("DSV: useEffect: gameRecords:", gameRecords);
+    console.log("DSV: useEffect: newGameRecordsByDay:", newGameRecordsByDay);
+    console.log("DSV: useEffect: largestKey:", largestKey.current);
+    console.log("DSV: useEffect: newDailyMax:", newDailyMax);
+
+    setDailyMax(newDailyMax);
+    setGameRecordsByDay(newGameRecordsByDay);
+  }, []);
+
   return (
     <div className="daily-streak-container">
-      {/* <div className="daily-streak-cell-00">7/3</div>
-      <div className="daily-streak-cell-10">7/4</div>
-      <div className="daily-streak-cell-20">7/5</div>
-      <div className="daily-streak-cell-30">7/6</div>
-      <div className="daily-streak-cell-40">7/7</div>
-      <div className="daily-streak-cell-50">7/8</div>
-      <div className="daily-streak-cell-60">7/9</div>
-      <div className="daily-streak-cell-70">7/10</div>
-      <div className="daily-streak-cell-80">7/11</div>
-      <div className="daily-streak-cell-90">7/12</div>
-      <div className="daily-streak-cell-100">7/13</div> */}
-
-      {/* Temp daily streak vis */}
-      {/* {
-        cellDates.map((cellDate, index) => (
-          <div className={getRandomClassName()} index={index}>{cellDate}</div> 
-        ))
-      } */}
-      
-      {/* Weeks header */}
-
-      <div className="daily-streak-week week-header">
-        {
-          weekdays.map((weekDay, index) => (
-            <div className="daily-streak-cell-info" key={index}>{weekDay}</div>
-          ))
-        }
-      </div>
-
       {
-        cellWeeks.map((week, index) => (
-          <div 
-            className="daily-streak-week"
-            key={index}
-          >
+        !gameRecordsByDay ? (
+          // TODO: Create better loading component
+          (<p>Loading...</p>)
+        ) : (
+          <>
+            {/* Weeks header */}
+
+            <div className="daily-streak-week week-header">
+              {
+                weekdays.map((weekDay, index) => (
+                  <div className="daily-streak-cell-info" key={index}>{weekDay}</div>
+                ))
+              }
+            </div>
+
             {
-              week.map((cellDate, index) => (
-                index === 0 ? (
-                  <div className="daily-streak-cell-info" key={index}>{cellDate + 1}</div>
-                ) : (
-                  <div className={`daily-streak-cell ${getRandomClassName()}`} key={index}>{cellDate}</div> 
-                )
+              cellWeeks.map((week, index) => (
+                <div 
+                  className="daily-streak-week"
+                  key={index}
+                >
+                  {
+                    week.map((cellDate, index) => (
+                      index === 0 ? (
+                        <div className="daily-streak-cell-info" key={index}>{cellDate + 1}</div>
+                      ) : (
+                        <div 
+                          className={`daily-streak-cell ${getCellClassName(cellDate)}`} 
+                          key={index}
+                        >
+                          <div className="daily-streak-cell-date">
+                            {cellDate}
+                          </div>
+
+                          <div className="daily-streak-cell-value">
+                            {getPuzzleCountOnDay(cellDate)}
+                          </div>
+                            
+                        </div> 
+                      )
+                    ))
+                  }
+                </div>
               ))
             }
-          </div>
-        ))
+          </>
+        )
       }
-
-      
     </div>
   );
 }
