+++ src/pages/MenuScreen/Menus/ProfileMenu.jsx
@@ -19,8 +19,159 @@ import DailyStreakVisualization from '../../../components/DailyStreakVisualizati
 import MenuHeader from '../../../components/MenuHeader/MenuHeader';
 import MenuLinks from '../../../components/MenuLinks/MenuLinks';
 
-// import convertMillisToMinutesAndSeconds from '../../../utils/ConvertMillisToMinutesAndSeconds';
 import convertMillisToMinutesAndSeconds from '../../../utils/convertMillisToMinutesAndSeconds';
+import convertFromFirestoreTimestampToDate from '../../../utils/convertFromFirestoreTimeStampToDate';
+
+// TODO: REMOVE, TEMP DATA
+const TEMP_GAME_RECORDS = [
+  {
+    completed: true,
+    lastPlayed: { seconds: 1688462821, nanoseconds: 0 },
+    gameTimer: 214500,
+    testGameRecord: true,
+    moves: 55,
+  },
+  {
+    completed: true,
+    lastPlayed: { seconds: 1688384400, nanoseconds: 0 },
+    gameTimer: 289000,
+    testGameRecord: true,
+    moves: 76,
+  },
+  {
+    completed: true,
+    lastPlayed: { seconds: 1688380800, nanoseconds: 0 },
+    gameTimer: 191000,
+    testGameRecord: true,
+    moves: 42,
+  },
+  {
+    completed: true,
+    lastPlayed: { seconds: 1688377200, nanoseconds: 0 },
+    gameTimer: 339000,
+    testGameRecord: true,
+    moves: 68,
+  },
+  {
+    completed: true,
+    lastPlayed: { seconds: 1688313600, nanoseconds: 0 },
+    gameTimer: 150000,
+    testGameRecord: true,
+    moves: 33,
+  },
+  {
+    completed: true,
+    lastPlayed: { seconds: 1688197200, nanoseconds: 0 },
+    gameTimer: 346000,
+    testGameRecord: true,
+    moves: 80,
+  },
+  {
+    completed: true,
+    lastPlayed: { seconds: 1688190000, nanoseconds: 0 },
+    gameTimer: 173000,
+    testGameRecord: true,
+    moves: 58,
+  },
+  {
+    completed: true,
+    lastPlayed: { seconds: 1687981200, nanoseconds: 0 },
+    gameTimer: 212000,
+    testGameRecord: true,
+    moves: 49,
+  },
+  {
+    completed: true,
+    lastPlayed: { seconds: 1687472400, nanoseconds: 0 },
+    gameTimer: 243000,
+    testGameRecord: true,
+    moves: 63,
+  },
+  {
+    completed: true,
+    lastPlayed: { seconds: 1687081200, nanoseconds: 0 },
+    gameTimer: 116000,
+    testGameRecord: true,
+    moves: 39,
+  },
+  {
+    completed: true,
+    lastPlayed: { seconds: 1686882000, nanoseconds: 0 },
+    gameTimer: 304000,
+    testGameRecord: true,
+    moves: 72,
+  },
+  {
+    completed: true,
+    lastPlayed: { seconds: 1686878400, nanoseconds: 0 },
+    gameTimer: 233000,
+    testGameRecord: true,
+    moves: 61,
+  },
+  {
+    completed: true,
+    lastPlayed: { seconds: 1686874800, nanoseconds: 0 },
+    gameTimer: 398000,
+    testGameRecord: true,
+    moves: 82,
+  },
+  {
+    completed: true,
+    lastPlayed: { seconds: 1686871200, nanoseconds: 0 },
+    gameTimer: 221000,
+    testGameRecord: true,
+    moves: 59,
+  },
+  {
+    completed: true,
+    lastPlayed: { seconds: 1686788400, nanoseconds: 0 },
+    gameTimer: 278000,
+    testGameRecord: true,
+    moves: 77,
+  },
+  {
+    completed: true,
+    lastPlayed: { seconds: 1686692400, nanoseconds: 0 },
+    gameTimer: 186000,
+    testGameRecord: true,
+    moves: 40,
+  },
+  {
+    completed: true,
+    lastPlayed: { seconds: 1686596400, nanoseconds: 0 },
+    gameTimer: 317000,
+    testGameRecord: true,
+    moves: 70,
+  },
+  {
+    completed: true,
+    lastPlayed: { seconds: 1686500400, nanoseconds: 0 },
+    gameTimer: 222000,
+    testGameRecord: true,
+    moves: 56,
+  },
+  {
+    completed: true,
+    lastPlayed: { seconds: 1686010800, nanoseconds: 0 },
+    gameTimer: 130000,
+    testGameRecord: true,
+    moves: 31,
+  },
+  {
+    completed: true,
+    lastPlayed: { seconds: 1684234800, nanoseconds: 0 },
+    gameTimer: 139800,
+    testGameRecord: true,
+    moves: 38,
+  },
+];
+
+// TODO: Also temp, but I might reuse this name later on
+//  this should also be a prop
+const gameRecords = TEMP_GAME_RECORDS.map((record) => {
+  record.lastPlayed = convertFromFirestoreTimestampToDate(record.lastPlayed);
+  return record;
+});
 
 const ProfileMenu = () => {
   const {
@@ -129,7 +280,7 @@ const ProfileMenu = () => {
             <h1>Daily Streak</h1>
 
             <DailyStreakVisualization
-              
+              gameRecords={gameRecords}
             />
 
             <p><strong>TODO:</strong> Add "Longest Unbroken Streak"</p>
