+++ src/styles/globals.scss
@@ -104,7 +104,10 @@ $button-hover-text-color: $dark-gray-300;
 $content-container-completed-color: $dark-gray-300;
 $content-container-completed-background-color: $light-gray-100;
 $content-container-completed-border-color: $light-gray-100;
-$daily-streak-cell-color: $light-gray-400;
+$daily-streak-cell-color: $light-gray-100;
+$daily-streak-cell-inverted-color: $dark-gray-100;
+$daily-streak-cell-empty-color: color.adjust($light-gray-400, $alpha: -.9);
+$daily-streak-cell-hover-color: $light-gray-100;
 $delete-unlocked-fill-color: $red-300;
 $delete-unlocked-font-color: $light-gray-100;
 $error-message-background-color: $red-100;
